fix: change focus behavior on the page
