+++ virtual-keyboard/index.html
@@ -30,7 +30,7 @@
 
 <body>
 
-  <script src="js/script.js"></script>
+  <script src="js/bundle.js"></script>
 </body>
 
 </html>
