+++ virtual-keyboard/styles/style.css
@@ -151,6 +151,7 @@ body {
 /*-----Keyboard------------------------------------------------*/
 /*-------------------------------------------------------------*/
 textarea {
+  resize: none;
   border: 1px solid black;
   width: 100%;
   height: 200px;
@@ -170,9 +171,9 @@ textarea {
   -webkit-box-direction: normal;
       -ms-flex-direction: column;
           flex-direction: column;
-  gap: 15px 15px;
+  gap: 10px 10px;
   background-color: #c4dfe6;
-  padding: 20px 20px;
+  padding: 15px 15px;
   border-radius: 2px;
 }
 
@@ -209,5 +210,5 @@ textarea {
 }
 
 .Space {
-  min-width: 300px;
+  min-width: 390px;
 }
