+++ virtual-keyboard/styles/style.scss
@@ -57,6 +57,7 @@ body {
 /*-----Keyboard------------------------------------------------*/
 /*-------------------------------------------------------------*/
 textarea {
+  resize: none;
   border: 1px solid black;
   width: 100%;
   height: 200px;
@@ -72,9 +73,9 @@ textarea {
   flex-direction: column;
   // justify-content: center;
   // margin: 0 auto;
-  gap: 15px 15px;
+  gap: 10px 10px;
   background-color: #c4dfe6;
-  padding: 20px 20px;
+  padding: 15px 15px;
   border-radius: 2px;
 }
 
@@ -107,7 +108,7 @@ textarea {
 }
 
 .Space {
-  min-width: 300px;
+  min-width: 390px;
 }
 
 .ShiftLeft {
